Rename: 컴포넌트 폴더명 PascalCase로 변경
